refactor(studio): lift the start-option card into a shared component
The fileset page's "how do you want to start?" tiles are about to be reused by
the customization page. Move the card and its option type out of
CreateFilesetStart into components/StartOptions so both pages render the same
thing, and make the id a type parameter so each page keeps its own closed set of
options rather than sharing one union.

No visual change to the fileset page.
