+++ web/packages/studio/src/components/StartOptions/types.ts
@@ -1,0 +1,42 @@
+// SPDX-FileCopyrightText: Copyright (c) 2025-2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+// SPDX-License-Identifier: Apache-2.0
+
+import type { BadgeProps } from '@nvidia/foundations-react-core';
+import type { LucideIcon } from 'lucide-react';
+
+export interface StartOptionTag {
+  label: string;
+  color: NonNullable<BadgeProps['color']>;
+  kind: NonNullable<BadgeProps['kind']>;
+}
+
+/**
+ * One tile in a "How do you want to start?" row. Generic over the id so each create
+ * flow can pin its own union of entry points while sharing the card that renders them.
+ */
+export interface StartOption<Id extends string = string> {
+  id: Id;
+  title: string;
+  description: string;
+  icon: LucideIcon;
+  tag?: StartOptionTag;
+  /**
+   * Whether this option is wired up. Disabled options still render (so the full set
+   * of future entry points is visible) but are no-ops — they cannot be selected and
+   * never reveal a detail panel or the Continue footer.
+   */
+  enabled: boolean;
+}
+
+export interface StartOptionCardProps {
+  option: StartOption;
+  selected: boolean;
+  /** Fired on click / keyboard activation. Only invoked for enabled options. */
+  onSelect: () => void;
+  /**
+   * Lays the icon beside the title instead of above it and drops the fixed height, so the
+   * row takes about half the vertical space. For flows where the tiles are a step on the
+   * way somewhere rather than the main event.
+   */
+  compact?: boolean;
+}
